requirements.txt
pynput==1.7.6
requests>=2.32.2,<3
Pillow==10.0.0
